refactor(Tooltip): Move to `partials`
Consistent with existing patterns
