+++ packages/react-component-library/src/components/Tooltip/partials/StyledMessage.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { color, fontSize } = selectors
+
+export const StyledMessage = styled.div`
+  font-size: ${fontSize('base')};
+  color: ${color('neutral', '100')};
+`
