+++ packages/react-component-library/src/components/Tooltip/partials/StyledTitle.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing } = selectors
+
+export const StyledTitle = styled.div`
+  padding-bottom: ${spacing('1')};
+`
