+++ packages/react-component-library/src/components/Tooltip/partials/StyledTooltip.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledTooltip = styled.div`
+  position: absolute;
+`
